Rename result class
