+++ stage/analysis_stage.py
@@ -1,5 +1,5 @@
 """Analyzing the evaluation output of the fuzz target and build script."""
-from result_classes import Result
+from results import Result
 from stage.base_stage import BaseStage
